Added comments, added multiple scripts, and fixed several bugs
Corrected the position of connection corners when dragging those with many corners. Added scripts for exporting attribute data, visualizing current attack paths, and disabling defenses. Fixed highlighting when zooming not following the corresponding block. Fixed bug that directional connections where disconnected when pressed but not moved. Fixed bug that the line of directional connections would not move with the triangle blocks when the start block was not connected